Moved formik date-picker to main.tsx as this is global configuration (and making the examples lazy would otherwise give errors)
