+++ packages/examples-formik/src/main.tsx
@@ -1,10 +1,32 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import { config as niceFormConfig } from '@ebay/nice-form-react';
+import NiceForm, { config as niceFormConfig } from '@ebay/nice-form-react';
 import formikAdapter from '@ebay/nice-form-react/adapters/formikAdapter';
 import formikMuiAdapter from '@ebay/nice-form-react/adapters/formikMuiAdapter';
+import { DatePicker } from '@mui/x-date-pickers/DatePicker';
+import type { Dayjs } from 'dayjs';
+import { fieldToDatePicker } from 'formik-mui-x-date-pickers';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import './index.css';
+
+const MyDatePicker = ({ ...props }) => {
+  props.onChange = (value: Dayjs | null) => {
+    props.form.setFieldTouched(props.field.name, true, false);
+    props.form.setFieldValue(props.field.name, value, true);
+    props.field.onChange(value);
+  };
+  props.onBlur = () => {
+    props.form.setFieldTouched(props.field.name, true, true);
+    props.field.onBlur();
+  };
+  return (
+    <DatePicker {...fieldToDatePicker({ field: props.field, form: props.form, meta: props.meta })}>
+      {props.children}
+    </DatePicker>
+  );
+};
+
+NiceForm.defineWidget('date-picker', MyDatePicker, ({ field }) => field);
 
 niceFormConfig.addAdapter(formikAdapter);
 niceFormConfig.addAdapter(formikMuiAdapter);
